zero leaks but 1559 errors

+++ logic/get_next_line.h
@@ -80,7 +80,7 @@ typedef struct s_game
 #include "./floodfill.c"
 
 
-int	ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strdup(const char *s);
 char	*ft_strchr(const char *s, int i);
